requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-core
botbuilder-schema
pandas
python-dotenv
pandasai
